feat(login): refactor login components to use jsx file extensions
